% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add}
\alias{add}
\title{Add together two numbers.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
add(1, 1)
add(10, 1)
}
