% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateSeats.R
\name{allocateSeats}
\alias{allocateSeats}
\title{allocateSeats
Allocates seats based on the district magnitude (number of seats) and the electoral formula.}
\usage{
allocateSeats(voteTotalsAllocateSeats, altsDataFrameAllocateSeats,
  numSeatsAllocateSeats, electoralFormulaAllocateSeats)
}
\arguments{
\item{voteTotalsAllocateSeats}{A 1 x numberOfCompetitors data frame with the vote totals recieved by each competitor.}

\item{altsDataFrameAllocateSeats}{a voteR dataFrame that contains the competitorID or alternativeID, and current location of competitors or alternatives being voted on.}

\item{numSeatsAllocateSeats}{A scalar. The district magnitude (number of seats available).}

\item{electoralFormulaAllocateSeats}{A string indicating the electoral}
}
\value{
outSeatAllocation
}
\description{
Takes the vote totals calculated by \code{\link{votersVote}} and allocates seats 
based on the district magnitude (number of seats) and the electoral formula.
}
