% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahldata.R
\docType{data}
\name{btw_districtlevel}
\alias{btw_districtlevel}
\title{German "Bundestagswahlen" election results for the constituency level 1949-2017}
\format{A data frame with 10633 rows and 74 variables:
\describe{
  \item{year}{election year}
  \item{WKID}{constituency ID}
  \item{type}{first or second vote}
  \item{WKN}{constituency name}
  \item{WKÜ}{state affiliation}
  \item{wahlberechtigte}{registered voters}
  \item{waehler}{voters}
  \item{ungueltige}{invalid votes}
  \item{gueltige}{valid votes}
  \item{...}{other parties}
}}
\usage{
btw_districtlevel
}
\description{
German "Bundestagswahlen" election results for the constituency level 1949-2017
}
\keyword{datasets}
