% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahldata.R
\docType{data}
\name{bundestag_laenderebene}
\alias{bundestag_laenderebene}
\title{lala}
\format{A data frame with 236 rows and 13 variables:
\describe{
  \item{year}{election year}
  \item{land}{regional level}
  \item{date}{election date}
  \item{...}{party vote share vector}
  \item{wbt}{voter turnout}
  \item{others}{other parties}
  \item{level}{Level of Election (federal or regional)}
}}
\usage{
bundestag_laenderebene
}
\description{
Bundestag election results for the regional laender level 1949-2017
}
\keyword{datasets}
