% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_identical.R
\name{check_identical}
\alias{check_identical}
\title{Title: Check for identical rows in a data frame}
\usage{
check_identical(data, dup_num, logical, name = "count")
}
\arguments{
\item{data}{An object of class data frame}

\item{dup_num}{A numeric value indicating the number of identical rows to check for}

\item{logical}{A character value indicating the logical operator to use. Must be one of 'greater', 'equal', or 'less'}

\item{name}{A character value indicating the name of the column to create in the output data frame. Default is 'count'}
}
\value{
A numeric value indicating the number of identical rows in the data frame
}
\description{
This function takes as an input tabular data with variables corresponding to ballot-level voting patterns and returns the number of identical rows in the data frame. The purpose is to identify the number of permutations for which there are $n$ number of duplicates.
}
\examples{
data <- data.frame(a = c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5),
                  b = c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5))

check_identical(data, 2, "greater")
}
