% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitorsAdapt.R
\name{competitorsAdapt}
\alias{competitorsAdapt}
\title{competitorsAdapt
Competitors Adapt According to their type/}
\usage{
competitorsAdapt(competitorDataFrameCompetitorAdapts,
  voteTotalsForCompetitorsOldCompetitorAdapts,
  voteTotalsForCompetitorsNewCompetitorAdapts,
  seatAllocationForCompetitorsNewCompetitorAdapts,
  seatAllocationForCompetitorsOldCompetitorAdapts,
  electoralFormulaCompetitorAdapts,
  scaledCompetitorAdaptStepSizeCompetitorAdapts,
  dimOneBoundsCompetitorsAdapt, dimTwoBoundsCompetitorsAdapt)
}
\arguments{
\item{competitorDataFrameCompetitorAdapts}{a voteR dataFrame that contains the competitorID, current location, and comeptitor type for each competitor.}

\item{voteTotalsForCompetitorsOldCompetitorAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the previous period.}

\item{voteTotalsForCompetitorsNewCompetitorAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the current period.}

\item{seatAllocationForCompetitorsNewCompetitorAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the previous period.}

\item{seatAllocationForCompetitorsOldCompetitorAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the current period.}

\item{electoralFormulaCompetitorAdapts}{The electoral formula being used in the model. In Plurality competitors maximize vote share, in PR systems they maximize seats.}

\item{scaledCompetitorAdaptStepSizeCompetitorAdapts}{Size step the party can take. Should be in cartesian distance units.}

\item{dimOneBoundsCompetitorsAdapt}{The bounds of the first dimension.}

\item{dimTwoBoundsCompetitorsAdapt}{The bounds of the second dimension.}
}
\value{
outCompetitorDataFrameCompetitorAdapts
}
\description{
Takes a voter data frame, a competitor position data frame, the voters utility values for all competitors over the past two time periods and the voteTotals of all competitors from the past 2 time periods and applies the appropriate adaptation rule for each competitor based on their competitor type.
}
