% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partynames.R
\name{distance_function}
\alias{distance_function}
\title{Calculate Distances}
\usage{
distance_function(data_in = "gles2017_out", who = "schwarzgelb",
  issue = "soz")
}
\value{
The vector including all coalitions.
}
\description{
Calculate Distances from different parties/koalitions
}
\examples{
distance_function(data_in = "gles2017_out",
                  who = "schwarzgelb",
                  issue = "soz")
}
