% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distToMargMed.R
\name{distToMargMed}
\alias{distToMargMed}
\title{distToMargMed
Finds the distance between a set of points and the marginal median.}
\usage{
distToMargMed(competitorDataFrameDistToMargMed, marginalMedianDistToMargMedian)
}
\arguments{
\item{competitorDataFrameDistToMargMed}{competitor data frame}

\item{marginalMedianDistToMargMedian}{the marginal median}
}
\value{
outDistanceToMarginalMedians
}
\description{
The marginal median is the vector of unidimensional medians.
}
