% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findICPoints}
\alias{findICPoints}
\title{findICPoints}
\usage{
findICPoints(voterID, idealPoint, altPointVector, orderScalar, salienceVector,
  precision = 0.01)
}
\arguments{
\item{idealPoint}{An ideal point for a voter.}

\item{altPointVector}{An alternative in the choice set.}

\item{orderScalar}{Order for the voter's distance metric. 1 is the Manhattan
Distance, 2 is the Cartesian Distance and 100000 is the Chebechev
Distance. You can choose any integer inbetween.}

\item{salienceVector}{The salience the voter has for each issue dimension. The element that
coresponds to the least salient matrix is a 1 and all the other dimensions have their salience measured
relative to that dimension.}

\item{precision}{How precise you want the IC to be. The lower the precision the higher the number
points on the IC that are found.}
}
\value{
indifferencePoints A set of points that are on the indifference curve.
}
\description{
Takes an ideal point and and alternative and returns the requested number of
points that are on the indifference curve for the alternative. The function
can take into account different distance metrics, and the
salience voters have over the different issue dimensions.
}
\details{
Possible TODO: Could add a simplify from <boost/geometry/algorithms/simplify.hpp> to possibaly decrease the number of points in each IC.
}
