% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findICsForSetOfVoters.R
\name{findICsForSetOfVoters}
\alias{findICsForSetOfVoters}
\title{findICsForSetOfVoters
Finds the indifference curves for the set of voterDataFrames provided.}
\usage{
findICsForSetOfVoters(votersDataFrame, altPoint, precision)
}
\arguments{
\item{votersDataFrame}{The voters data frame must have a specific format, and it must be an R data.frame object. There must be at least these 6 variables and they must have the following names. The order of the variables in the data.frame is not important as long as the variables have the proper names.

 voterID: A numeric identifier unique to the voter.
 xIdeal: The x coordinate of the voter's ideal point.
 yIdeal: The y coordinate of the voter's ideal point.
 minkoOrder: The Minkowski order of the voters MInkowski metric based utility function. = 1, is City Block. = 2 is Euclidian and 100 = is  See ?Minkowski. 
 xSalience: The salience of the x dimension for the voter. The dimension with the lowest salience is normalized to 1 and it is the numerarier, the salience of other dimension is measured in units of the numerarire. 
 ySalience: The salience of the y dimension for the voter. he dimension with the lowest salience is normalized to 1 and it is the numerarier, the salience of other dimension is measured in units of the numerarire.}

\item{altPoint}{The alternative that the voter's are evaluating. The indifference curves are the set of all points that gives the voter the same utility as this point.}

\item{percision}{How precisely you want to trace out the voters' indifference curves. The lower the number the more accurate and smooth the indifference curve will be, but for large numbers of voters larger numbers will slow down plotting. Default: precision = .01}
}
\value{
A list where each element of the list contains the coordinantes of the voter's indfference curve.
}
\description{
findICsForSetOfVoters
Finds the indifference curves for the set of voterDataFrames provided.
}
