% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findParetoSet}
\alias{findParetoSet}
\title{findParetoSet}
\usage{
findParetoSet(idealPoints)
}
\arguments{
\item{idealPoints}{A numberOfVoters x numberOfDimensions matrix of ideal points.}
}
\value{
The set of points that constitutes the pareto set of the set of alternatives. The points are listed down the rows in clockwise
or counterclockwise order with the first and last element in the matrix being the same ideal point.
}
\description{
Will find the pareto set in a two dimensional spatial voting model given a set of ideal points.
}
\details{
The pareto set of a set of ideal points is the convex hull of that set of
alternatives. This function will take a set of ideal points and returns
their convex hull as a matrix of points. The points are listed down the rows in clockwise
or counterclockwise order with the first and last element in the matrix being the same ideal point.
}
