% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findSuperElipseRadius}
\alias{findSuperElipseRadius}
\title{superElipseRadius}
\usage{
findSuperElipseRadius(idealPoint, altPoint, orderScalar, salienceVector)
}
\arguments{
\item{idealPoint}{An ideal point for a voter.}

\item{altPoint}{An alternative in the choice set.}

\item{orderScalar}{Order for the voter's distance metric. 1 is the Manhattan
Distance, 2 is the Cartesian Distance and 100000 is the Chebechev
Distance. You can choose any integer inbetween.}

\item{salienceVector}{The salience the voter has for each issue dimension. The element that
coresponds to the least salient matrix is a 1 and all the other dimensions have their salience measured
relative to that dimension.}
}
\value{
indifferencePoints A set of points that are on the indifference curve.
}
\description{
Returns the distance between the center of a super elipse and a point on the
elipse. The distance is mesured in the "order" of the superelipse. This
order is the same as the minkowski order for distance.
}
