% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findWinSet}
\alias{findWinSet}
\title{findWinSet}
\usage{
findWinSet(indifferenceCurves)
}
\arguments{
\item{indifferenceCurves}{A numberOffVoters list of data frames from R. Each data frame is the point set for a voters}

\item{kToDefeatSQ}{A scalar indicating the number of indifference curves that must overlap to create the Win Set 
(aka the k-majority rule).}
}
\value{
winSetList A list of data frames where each data frame is a point set that defines one of the disjoint sets that are the 
    boundries of the win set. If there are less than two disjoint sets then the list will contain only one data frame.
}
\description{
Takes a set of indifference curves and returns the k-najority Rule WinSet. The IC's should be stored as a list
of DataFrames where each data framd in the list is the point set that constitutes the boundries of each voter's Indifference Curve. 
These ICs are usually generated by \code{\link{findICPoints}}
}
\details{
Possible TODO: 1) Could add a simplify at the begining from <boost/geometry/algorithms/simplify.hpp> to possibaly decrease the number of points in each IC.
                2) Also could switch entire package to use Well Know Text as is used in the {sf} package. Then all polygons could be 
                supplied as a single elements in a list (rather than as data frames in a list.)
}
