% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCompetitorHeading.R
\name{genCompetitorHeading}
\alias{genCompetitorHeading}
\title{genCompetitorHeading
Assigns each competitor a heading in radians.}
\usage{
genCompetitorHeading(numCompetitorsGenCompHeading)
}
\arguments{
\item{numCompetitorsGenCompHeading}{scalar Number of voters}
}
\value{
outCompetitorType numCompetitors x 1 A vector of competitor types, one for each competitor.
}
\description{
Gives an angle in radians for each competitor. For competitors with types that move this is their inital heading. 
\code{\link{genCompetitorHeading}} is used in conjuction 
with \code{\link{genCompetitorHeading}} and\code{\link{genIdeals}} by \code{\link{genCompetitors}} to generate
the characteristics of each competitor.
}
