% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genIdeals.R
\name{genIdeals}
\alias{genIdeals}
\title{genIdeals
Generates ideal points for a set of voters.
 
Generates one or two dimensional ideal points from one of given set of distributions, from calibrated data, or from a user provided distribution. 
it is a wrapper for \code{\link{genPoints}} and is itself further wrapped by \code{\link{genVoters}}.}
\usage{
genIdeals(
  numberOfDimensionsGenIdeals = 1,
  numberOfIdealsGenIdeals = 100,
  distributionTypeGenIdeals = "unif",
  distributionParametersGenIdeals = c(-1, 1),
  dimOneBoundsGenIdeals = c(-Inf, Inf),
  dimTwoBoundsGenIdeals = c(-Inf, Inf)
)
}
\arguments{
\item{numberOfDimensionsGenIdeals}{The number of policy dimensions.}

\item{numberOfIdealsGenIdeals}{Number of points to generate.}

\item{distributionTypeGenIdeals}{A string identifying the base R discribution to draw
the ideal points from. Uses the base R random number generation family of
commands rxxxx (see ?distributions). The user should specify the
distribution as a string using the standard R abreviation for the
distribution (see ?distributions for a list). Currently supported are: "norm",
"unif", "beta", "cauchy", "chisq", "weibull"}

\item{distributionParametersGenIdeals}{A vector that contains the additional
parameters needed by the particular rxxxx function for a distribtuion. (see
?rxxxx where xxxx is a function listed under ?distribution). Example for a
Normal(0,1), use: c(0,1).}

\item{dimOneBoundsGenIdeals}{A vector that contains the starting and ending poitns of t
he first dimension. Example: c(0,1). Defaults to c(-Inf, Inf) if no boundary is provided.}

\item{dimTwoBoundsGenIdeals}{A vector that contains the starting and ending poitns of t
he first dimension. Example: c(0,1). Defaults to c(-Inf, Inf)  if no boundary is provided.}
}
\value{
outIdeals An ideal point matrix that is numVoters x numDimensions
}
\description{
genIdeals
Generates ideal points for a set of voters.
 
Generates one or two dimensional ideal points from one of given set of distributions, from calibrated data, or from a user provided distribution. 
it is a wrapper for \code{\link{genPoints}} and is itself further wrapped by \code{\link{genVoters}}.
}
\examples{
  genIdeals(numberOfDimensionsGenIdeals=1, numberOfIdealsGenIdeals=100, distributionTypeGenIdeals ="norm", distributionParametersGenIdeals = c(0,.5), dimOneBoundsGenIdeals = c(0,Inf), dimTwoBoundsGenIdeals = c(-Inf,0))
  
  genIdeals(numberOfDimensionsGenIdeals=2, numberOfIdealsGenIdeals=100, distributionTypeGenIdeals ="unif", distributionParametersGenIdeals = c(-1,1), dimOneBoundsGenIdeals = c(-Inf,Inf), dimTwoBoundsGenIdeals = c(-Inf,Inf) )
  
 genIdeals(numberOfDimensionsGenIdeals=2, numberOfIdealsGenIdeals=100, distributionTypeGenIdeals ="beta", distributionParametersGenIdeals = c(.1,1), dimOneBoundsGenIdeals = c(-Inf,Inf), dimTwoBoundsGenIdeals = c(-Inf,Inf) )
  
}
