% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLossOrder.R
\name{genLossOrder}
\alias{genLossOrder}
\title{genLossOrder 
Generates Loss Function Orders for voters in a generated voter set.}
\usage{
genLossOrder(
  numVotersGenLoss,
  lossOrderForAllGenLoss = FALSE,
  probabilityLinearGenLoss = 1,
  probabilityQuadraticGenLoss = 0,
  probabilityCubicGenLoss = 0
)
}
\arguments{
\item{numVotersGenLoss}{scalar Number of voters}

\item{lossOrderForAllGenLoss}{logical or scalar If If FALSE (the default) then the agent's loss order for their utility function is determined by the other parameters in this function. If a number is provided this will be the order of the loss function for all voters.}

\item{probabilityLinearGenLoss}{scalar [0,1] The probability that any given voter will have elliptical (Minkowski Order of 2) indifference curves.}

\item{probabilityQuadraticGenLoss}{scalar [0,1] The probability that any given voter will have Manhattan/Diamond (Minkowski Order = 1) indifference curves.}

\item{probabilityCubicGenLoss}{scalar [0,1] The probability that any given voter will have Chebyshev/Square (Minkowski Order = 100) indifference curves.}
}
\value{
outMinkowskiOrder numVoters x 1 A vector of minkowski orders, one for each voter that determines the overall shape of the voter's indifference curves.
}
\description{
Randomly generate a Loss Function Order for all voters in a data set. \code{\link{genLossOrder}} is used in conjuction 
with \code{\link{GenMinkoOrder}},  \code{\link{genSalience}} and \code{\link{genIdeals}} by \code{\link{genVoters}} to generate
the parameters of each agent's Salience Weighted Minkowski Utility function.
}
