% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMinkoOrder.R
\name{genMinkoOrder}
\alias{genMinkoOrder}
\title{genMinkoOrder 
Generates Minkowski Orders for voters in a generated voter set.}
\usage{
genMinkoOrder(numVotersGenMinko, allEllipticalGenMinko = FALSE,
  probabilityElipticalGenMinko = 1, probabilityDiamondGenMinko = 0,
  probabilitySquareGenMinko = 0)
}
\arguments{
\item{numVotersGenMinko}{scalar Number of voters}

\item{allEllipticalGenMinko}{logical If If FALSE (the default) then the agent's Minkowski order for their utility function is determined by the other parameters in this function. If TRUE then all agents will have eliptical indifference curves (Minkowski Order = 2). If their salience's across all dimesnions is 1 then the indifference curves will be circles.}

\item{probabilityElipticalGenMinko}{scalar [0,1] The probability that any given voter will have elliptical (Minkowski Order of 2) indifference curves.}

\item{probabilityDiamondGenMinko}{scalar [0,1] The probability that any given voter will have Manhattan/Diamond (Minkowski Order = 1) indifference curves.}

\item{probabilitySquareGenMinko}{scalar [0,1] The probability that any given voter will have Chebyshev/Square (Minkowski Order = 100) indifference curves.}
}
\value{
outMinkowskiOrder numVoters x 1 A vector of minkowski orders, one for each voter that determines the overall shape of the voter's indifference curves.
}
\description{
Randomly generate a MinkowskiOrder for all voters in a data set. \code{\link{genMinkoOrder}} is used in conjuction 
with \code{\link{genSalience}} and \code{\link{genIdeals}} by \code{\link{genVoters}} to generate
the parameters of each agent's Salience Weighted Minkowski Utility function.
}
