% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSincereProposals.R
\name{genSincereProposals}
\alias{genSincereProposals}
\title{GenSincereProposals
Generates a set of sincere proposals to face the status quo policy.}
\usage{
genSincereProposals(
  voterIdealsGenSincereProposals,
  numberOfSincereProposalsGenSincereProposals = 1
)
}
\arguments{
\item{voterIdealsGenSincereProposals}{The dataframe of voter ideals usually created with \code{\link{genVoters}}. 
If not creaeted with \code{\link{genVoters}} then the user must make sure the coordinates for the ideal points
in the sumbitted dataframe are named xLocation and yLocation.}

\item{numberOfSincereProposalsGenSincereProposals}{Number of sincere proposals to generate. Default is 1.}
}
\value{
outSincereProposalsDataFrame data.frame The RandomProposals data frame will have the following format.
 proposalID: A numeric identifier unique to the voter.
 xLocation: The x coordinate of the voter's ideal point.
 yLocation: The y coordinate of the voter's ideal point.
}
\description{
Generates a set of sincere proposals to face the status quo policy. 
 For the number of porposals requested, this fucntion will at random select that many voter's
 ideal points as the proposals to face the status quo.
}
\examples{
   someVoters <- genVoters(numberOfDimensionsGenVoters = 2) 
  genSincereProposals(voterIdealsGenSincereProposals = someVoters, numberOfSincereProposalsGenSincereProposals = 2) 
  
  genSincereProposals(voterIdealsGenSincereProposals = someVoters) 
  
}
