% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gles.R
\name{gles_recode_partyvar}
\alias{gles_recode_partyvar}
\title{gles_recode_partyvar}
\usage{
gles_recode_partyvar(year = 2017, dataset_input = "gles2017",
  dataset_output = "gles2017_out", varname = "q52", own = NULL,
  varlabel = "soz", key = c("a", "b", "c", "d", "e", "f", "g"),
  partynames = c("cdu", "csu", "spd", "linke", "gruene", "fdp", "afd"),
  NAs = "<0", plot = TRUE)
}
\arguments{
\item{year}{year the GLES-study is from. Defaults to 2017.}

\item{dataset_input}{Character string of the name of a dataframe containing the raw data.}

\item{dataset_output}{Character string of the name the output data frame (may already exist or not).}

\item{varname}{Character string of the name of the original variable.}

\item{own}{May apply: Different variable name for own position (on left-right scales, e.g.)}

\item{varlabel}{Character string of the to-be-assigned variable label.}

\item{key}{Character vector containing original alphabetic party keys.}

\item{partynames}{Character vector containing shortname party keys.}

\item{NAs}{Numeric vector containing to-be-assigned NAs/Missing values.}

\item{plot}{Logical T/F: Show relative frequency barplots while plotting.}
}
\value{
A data frame containing output dataframe including newly appended new-variables.
}
\description{
Recode a multiparty-variable in GESIS-Datasets such as the GERMAN LONGITUDINAL ELECTION STUDY (GLES)
}
\examples{
gles_recode_partyvar(year = 2017,
                     dataset_input = "gles2017",
                     dataset_output = "gles2017_out",
                     varname = "q52",
                     varlabel = "soz",
                     key = c("a","b","c","d","e","f","g"),
                     partynames = c("cdu","csu","spd","linke","gruene","fdp","afd"),
                     NAs = c(-97,-98,-99))
}
