% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Function %nin%}
\usage{
x \%nin\% y
}
\value{
Logical Vector
}
\description{
The opposite of %in%
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
c(2,4,8) \%nin\% c(1,4,7)
}
