% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunterNewHeading.R
\name{hunterNewHeading}
\alias{hunterNewHeading}
\title{hunterNewHeading.R
For the Hunter type competitor. Takes a heading in radians and randomly draws a new heading between pi/2 to 3pi/2 (90 to 270 degrees).}
\usage{
hunterNewHeading(origHeadingInRadians)
}
\arguments{
\item{origHeadingInRadians}{The inital point in 2D space.}
}
\value{
outNewHeading
}
\description{
hunterNewHeading.R
For the Hunter type competitor. Takes a heading in radians and randomly draws a new heading between pi/2 to 3pi/2 (90 to 270 degrees).
}
