% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntersAdapt.R
\name{huntersAdapt}
\alias{huntersAdapt}
\title{huntersAdapt
Hunters Adaptation rule}
\usage{
huntersAdapt(competitorDataFrameHunterAdapts,
  voteTotalsForCompetitorsOldHunterAdapts,
  voteTotalsForCompetitorsNewHunterAdapts,
  seatAllocationForCompetitorsNewHunterAdapts,
  seatAllocationForCompetitorsOldHunterAdapts, electoralFormulaHunterAdapts,
  scaledCompetitorAdaptStepSizeHunterAdapts, dimOneBoundsHuntersAdapt,
  dimTwoBoundsHuntersAdapt)
}
\arguments{
\item{competitorDataFrameHunterAdapts}{a voteR dataFrame that contains the competitorID, current location, and comeptitor type for each competitor.}

\item{voteTotalsForCompetitorsOldHunterAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the previous period.}

\item{voteTotalsForCompetitorsNewHunterAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the current period.}

\item{seatAllocationForCompetitorsNewHunterAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the previous period.}

\item{seatAllocationForCompetitorsOldHunterAdapts}{a numCOmpetitors x 2 matrix containing each competitor's competitorID and their vote share from the current period.}

\item{electoralFormulaHunterAdapts}{The electoral formula being used in the model. In Plurality competitors maximize vote share, in PR systems they maximize seats.}

\item{scaledCompetitorAdaptStepSizeHunterAdapts}{Size step the party can take. Should be in cartesian distance units.}

\item{dimOneBoundsHuntersAdapt}{The bounds of the first dimension.}

\item{dimTwoBoundsHuntersAdapt}{The bounds of the second dimension.}
}
\value{
outCompetitorDataFrameCompetitorAdapts
}
\description{
Hunters compare their vote share or seat share in an election at t, to their vote share or seat share in the election at t-1.
If the vote share/ seat share improved they continue moving in the same direction.
If the vote share/ seat share did not improve they randomly chooses a new heading between 90 to 270 degrees away from its previous heading.
}
