% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_run}
\alias{import_run}
\title{Import Run}
\usage{
import_run(file = "data/2017-11-06.gpx", layer = "track_points",
  type = "gpx", track_progress = FALSE)
}
\arguments{
\item{file}{path and file name.}

\item{layer}{Which layer you want to select. Use "track_points" for now.}

\item{type}{Type of log-file: currently, only gpx is supported.}

\item{track_progress}{T/F if data loading progress tracking is wished for}
}
\value{
A data frame data about XXXXXXX.
}
\description{
Import Run (Standard: GPX) to Data Frame
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
run <- import_run(file = 'data/2017-11-06.gpx',type = 'gpx')
head(run)
}
