% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gles.R
\name{intrakoadistanz}
\alias{intrakoadistanz}
\title{Calculate Intra-Coalition Heterogeneity}
\usage{
intrakoadistanz(who = "schwarzgelb", issue = "lr", input = "gles2017_out",
  year = 2017)
}
\value{
The vector including all coalitions.
}
\description{
Calculate Intra-Coalition Heterogeneity from different parties/koalitions
}
\examples{
distance_function(input = "gles2017_out",
                  who = "schwarzgelb",
                  issue = "soz")
}
