% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInParetoSetForVoterAndAltsDataFrame.R
\name{isInParetoSetForVoterAndAltsDataFrame}
\alias{isInParetoSetForVoterAndAltsDataFrame}
\title{isInParetoSetForVoterAndAltsDataFrame
Finds the indifference curves for the set of voterDataFrames provided.}
\usage{
isInParetoSetForVoterAndAltsDataFrame(alternativesDataFrame, votersDataFrame)
}
\arguments{
\item{votersDataFrame}{The voters data frame must have a specific format, and it must be an R data.frame object. There must be at least these 6 variables and they must have the following names. The order of the variables in the data.frame is not important as long as the variables have the proper names.

 ID: A numeric identifier unique to the voter.
 xLocation: The x coordinate of the voter's ideal point.
 yLocation: The y coordinate of the voter's ideal point.
 minkoOrder: The Minkowski order of the voters MInkowski metric based utility function. = 1, is City Block. = 2 is Euclidian and 100 = is  See ?Minkowski. 
 xSalience: The salience of the x dimension for the voter. The dimension with the lowest salience is normalized to 1 and it is the numerarier, the salience of other dimension is measured in units of the numerarire. 
 ySalience: The salience of the y dimension for the voter. he dimension with the lowest salience is normalized to 1 and it is the numerarier, the salience of other dimension is measured in units of the numerarire. 
 lossOrder: The loss order for the agents utility function. See the parameter lossOrderVector in ?minkowskiUtilitySets().

#' @param alternativesDataFrame The alternatives data frame must have a specific format, and it must be an R data.frame object. There must be at least these 3 variables and they must have the following names. The order of the variables in the data.frame is not important as long as the variables have the proper names.

 ID: A numeric identifier unique to the alternative.
 xLocation: The x coordinate of the alternative's location.
 yLocation: The y coordinate of the alternative's location.}

\item{alternativeToCheck}{An R vector that contains the corrdinates of the alternative that you want to check.}
}
\value{
DataFrame with two columns the first is the Alternatives ID and the second is a logical that is TRUE if the alt is in the pareto set and FALSE of the alt is not in the pareto set.
}
\description{
isInParetoSetForVoterAndAltsDataFrame
Finds the indifference curves for the set of voterDataFrames provided.
}
