% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isInParetoSetFromPointAndIdeals}
\alias{isInParetoSetFromPointAndIdeals}
\title{isInParetoSetFromPointAndIdeals}
\usage{
isInParetoSetFromPointAndIdeals(aSexpPoint, aSexpMatrix)
}
\arguments{
\item{aSexpPoint}{An R vector (SEXP) that contains the corrdinates of the
alternative that you want to check.}

\item{aSexpMatrix}{An R matrix (SEXP) that contains the ideal points of the
voters.}
}
\value{
Logical that is TRUE if the alt is in the pareto set and FALSE of
  the alt is not in the pareto set.
}
\description{
This function takes an alternative and a pareto set and checks to see if the
point is in the pareto set. The function \code{isInParetoSetFromPointAndPS()} 
will take an alternative and the set of points that define the pareto set and 
check if the alternative is in the pareto set.
}
