% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partynames.R
\name{koa_members}
\alias{koa_members}
\title{Get coalition members}
\usage{
koa_members(koalition)
}
\arguments{
\item{coalition}{Character string containing the name of the coalition. \cr
Options are c("jamaika","schwarzgelb","rotgruen","groko","rotrotgruen","ampel","schwarzgruen").}
}
\value{
A vector containing all parties included in the coalition.
}
\description{
Get parties that are member of a certrain coalition
}
\examples{
koa_members("schwarzgelb")
}
