% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahldata.R
\docType{data}
\name{laenderbip}
\alias{laenderbip}
\title{German regional GDP data}
\format{A data frame with 27 rows and 18 variables:
\describe{
  \item{year}{regional level}
  \item{...}{name of the bundesland}
  \item{d}{Germany total sum}
}}
\usage{
laenderbip
}
\description{
German regional GDP data
}
\keyword{datasets}
