% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahldata.R
\docType{data}
\name{landesregierungen}
\alias{landesregierungen}
\title{German regional government cabinets for the whole post-war period}
\format{A data frame with 325 rows and 11 variables:
\describe{
  \item{land}{regional level}
  \item{cabinet}{name of the respective cabinet}
  \item{years}{years cabinet was in office}
  \item{dates}{specific dates cabinet was in office}
  \item{parties}{parties forming the coalition}
  \item{primeminister}{name of the government leader}
  \item{party_x}{name of the forming parties}
  \item{start}{year in which gov't started}
  \item{end}{year in which it ended}
}}
\usage{
landesregierungen
}
\description{
German regional government cabinets for the whole post-war period
}
\keyword{datasets}
