% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahldata.R
\docType{data}
\name{landtagswahlen}
\alias{landtagswahlen}
\title{German "Landtagswahlen" election results for the regional laender level 1946-2018}
\format{A data frame with 233 rows and 13 variables:
\describe{
  \item{year}{election year}
  \item{land}{regional level}
  \item{date}{election date}
  \item{...}{party vote share vector}
  \item{wbt}{voter turnout}
  \item{others}{other parties}
  \item{level}{Level of Election (federal or regional)}
}}
\usage{
landtagswahlen
}
\description{
German "Landtagswahlen" election results for the regional laender level 1946-2018
}
\keyword{datasets}
