% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dictionary.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Title}
\usage{
make_dictionary(dataset, format = c("wide", "long"))
}
\arguments{
\item{dataset}{SPSS dataframe in tbl_df format imported with haven.}

\item{format}{specify whether you want a wide (each variable one row) or wide format (each value one row).}
}
\value{
a data.fram with meta information which can by querried with R-Studio functionalities.
}
\description{
Title
}
