% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveDistDAtAngleT.R
\name{moveDistDAtAngleT}
\alias{moveDistDAtAngleT}
\title{moveDistDAtAngleT
Takes a two dimensional coordinate and moves it to another coordinate based on a distance and angle.}
\usage{
moveDistDAtAngleT(origPointMoveDistDAtAngleT, distanceDMoveDistDAtAngleT,
  angleTMoveDistDAtAngleT, dimOneBoundsMoveDistDAtAngleT,
  dimTwoBoundsMoveDistDAtAngleT)
}
\arguments{
\item{origPointMoveDistDAtAngleT}{The inital point in 2D space.}

\item{distanceDMoveDistDAtAngleT}{The distance you want to move from the original point.}

\item{angleTMoveDistDAtAngleT}{The angle in radians you want to take when moving away from the original point.}

\item{dimOneBoundsMoveDistDAtAngleT}{Boundry points of the x dimension.}

\item{dimTwoBoundsMoveDistDAtAngleT}{Boundry poitns of the y dimension.}
}
\value{
outNewPoint
}
\description{
moveDistDAtAngleT
Takes a two dimensional coordinate and moves it to another coordinate based on a distance and angle.
}
