% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partynames.R
\name{parties}
\alias{parties}
\title{Get parties}
\usage{
parties(year = 2017)
}
\value{
The vector including all parties
}
\description{
Get main parties for Gles analysis
}
\examples{
parties(year = 2017)
}
