% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_poll}
\alias{plot_poll}
\title{Plot a multiparty poll}
\usage{
plot_poll(vote = c(cdu = 0.33, spd = 0.2, fdp = 0.11, linke = 0.09, gruene =
  0.09, afd = 0.12, sonstige = 0.05), order = "alphabetical",
  sample_confidence_bounds = TRUE, sample_n = 1000, n_draw = 10000,
  show_quantiles = c(0.05, 0.95), round = 1, xlab = "Party",
  ylab = "Voteshare", title = "Title", subtitle = "Subtitle",
  caption = "Caption", theme_ipsum = FALSE, grid = "Y")
}
\arguments{
\item{vote}{A labeled party vote share vector.}

\item{order}{Method to order parties (Default is "alphabetical"; also takes "descending" and "ascending" as well as manual specification of party vector)}

\item{sample_confidence_bounds}{Logical T/F: add empirical dirichlet quantiles}

\item{sample_n}{The number of observations in the poll sample.}

\item{n_draw}{How many samples to draw from the dirichlet distribution.}

\item{show_quantiles}{Vector of quantiles/confidence boundaries to calculate.}

\item{round}{Round to k decimals after comma}

\item{xlab}{x-label string}

\item{ylab}{y-label string}

\item{title}{title string}

\item{subtitle}{subtitle string}

\item{caption}{caption string}

\item{theme_ipsum}{Pre-applies nice theme from the hrbrthemes-package.(Attention: possible font-issues when Roboto font is not installed on your computer.)}

\item{grid}{(Applies only if theme_ipsum == T) Add a grid (options: "none","Y")}
}
\value{
A data frame containing \code{n} rows of samples for each party.
}
\description{
...
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
sample_dirichlet_quantiles(vote = c(cdu = 0.33,....
}
