% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_viz.R
\name{run2map}
\alias{run2map}
\title{Run2Map}
\usage{
run2map(run_track = import_run(file = "data/2017-11-06.gpx", layer =
  "tracks"))
}
\arguments{
\item{run}{a run data frame returned by \code{import_run}.}
}
\value{
A ggplot object.
}
\description{
Visualize Run2Map
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
run <- import_run(file = 'data/2017-11-06.gpx',type = 'gpx')
head(run)
}
