% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_viz.R
\name{run2profile}
\alias{run2profile}
\title{RunElecation}
\usage{
run2profile(run = import_run(), smooth = 2, title = "runR-Elevationplot",
  subtitle = "Untertitel", xlab = "Time", ylab = "Elevation (in meters)")
}
\arguments{
\item{run}{a run data frame returned by \code{import_run}.}
}
\value{
A ggplot object.
}
\description{
Visualize ...
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
run2profile(smooth = 0.5)
}
