% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polling.R
\name{sample_dirichlet}
\alias{sample_dirichlet}
\alias{sample_dirichlet}
\title{Dirichlet-sample of a multinomial election poll}
\usage{
sample_dirichlet(vote = c(cdu = 0.5, spd = 0.4, fdp = 0.1), sample_n = 1000,
  n_draw = 10000)

sample_dirichlet(vote = c(cdu = 0.5, spd = 0.4, fdp = 0.1), sample_n = 1000,
  n_draw = 10000)
}
\arguments{
\item{vote}{A labeled party vote share vector.}

\item{sample_n}{The number of observations in the poll sample.}

\item{n_draw}{How many samples to draw from the dirichlet distribution.}

\item{vote}{A labeled party vote share vector.}

\item{sample_n}{The number of observations in the poll sample.}

\item{n_draw}{How many samples to draw from the dirichlet distribution.}
}
\value{
A data frame containing \code{n} rows of samples for each party.

A data frame containing \code{n} rows of samples for each party.
}
\description{
Calculate a dirichlet-sample of a multinomial election poll

Calculate a dirichlet-sample of a multinomial election poll
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.


Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
sample_dirichlet(vote = c(cdu = 0.33,
                          spd = 0.20,
                          fdp = 0.11,
                          linke = 0.09,
                          gruene = 0.09,
                          afd = 0.12,
                          sonstige = 0.05),
                 sample_n = 1000,
                 n_draw = 10000)
sample_dirichlet(vote = c(cdu = 0.33,
                          spd = 0.20,
                          fdp = 0.11,
                          linke = 0.09,
                          gruene = 0.09,
                          afd = 0.12,
                          sonstige = 0.05),
                 sample_n = 1000,
                 n_draw = 10000)
}
