% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polling.R
\name{sample_dirichlet_quantiles}
\alias{sample_dirichlet_quantiles}
\alias{sample_dirichlet_quantiles}
\title{Empirical Dirichlet Quantiles from Multinomial Election Poll}
\usage{
sample_dirichlet_quantiles(vote = c(cdu = 0.5, spd = 0.4, fdp = 0.1),
  sample_n = 1000, n_draw = 10000, show_mean = TRUE,
  show_quantiles = c(0.1, 0.95), round = 2)

sample_dirichlet_quantiles(vote = c(cdu = 0.5, spd = 0.4, fdp = 0.1),
  sample_n = 1000, n_draw = 10000, show_mean = TRUE,
  show_quantiles = c(0.1, 0.95), round = 2)
}
\arguments{
\item{vote}{A labeled party vote share vector.}

\item{sample_n}{The number of observations in the poll sample.}

\item{n_draw}{How many samples to draw from the dirichlet distribution.}

\item{show_mean}{Logical T/F: Show sample mean.}

\item{show_quantiles}{Vector of quantiles/confidence boundaries to calculate.}

\item{round}{Logical T/F Round Results to k decimal digits.}

\item{show_mean}{Logical T/F: Show sample mean.}

\item{show_quantiles}{Vector of quantiles/confidence boundaries to calculate.}

\item{round}{Logical T/F Round Results to k decimal digits.}
}
\value{
A data frame containing \code{n} rows of samples for each party.

A data frame containing \code{n} rows of samples for each party.
}
\description{
Calculate empirical quantiles from a sample created by \code{\link{sample_dirichlet}} of a multinomial election poll

Calculate empirical quantiles from a sample created by \code{\link{sample_dirichlet}} of a multinomial election poll
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.


Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
sample_dirichlet_quantiles(vote = c(cdu = 0.33,
                          spd = 0.20,
                          fdp = 0.11,
                          linke = 0.09,
                          gruene = 0.09,
                          afd = 0.12,
                          sonstige = 0.05),
                 sample_n = 1000,
                 n_draw = 10000,
                 show_mean = TRUE,
                 show_quantiles = c(0.05,0.95))
sample_dirichlet_quantiles(vote = c(cdu = 0.33,
                          spd = 0.20,
                          fdp = 0.11,
                          linke = 0.09,
                          gruene = 0.09,
                          afd = 0.12,
                          sonstige = 0.05),
                 sample_n = 1000,
                 n_draw = 10000,
                 show_mean = TRUE,
                 show_quantiles = c(0.05,0.95))
}
