% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_wahlrecht.R
\name{scrape_wahlrecht_bund}
\alias{scrape_wahlrecht_bund}
\title{Scrape German Federal Level Polls from Wahlrecht.de}
\usage{
scrape_wahlrecht_bund(institutes = c("allensbach", "emnid", "forsa",
  "politbarometer", "gms", "dimap", "insa", "yougov"),
  include_rawdata = F, parties = c("cdu/csu", "spd", "grüne", "fdp",
  "linke", "afd"))
}
\arguments{
\item{institutes}{A vector with institutes to include.}

\item{include_rawdata}{Logical whether raw data (available for some instutues) shall be included. Raw polls can be identified via the raw_by_institue column. Defaults to FALSE.}

\item{parties}{A character vector of which parties to include. Defaults to the main parties currently represented in the German Bundestag.}
}
\value{
A data frame all polls since 1998 including Party Vote Shares along metadata including Year, Date of Publication, Start and End of Data Collection, Polling Institute, Sample Size and a "Raw"-Dummy.
}
\description{
Scrape German Federal Election Polls from the website "Wahlrecht.de"
}
\section{Notes}{

The "other parties" column is calculated as the difference of the sum of all selected parties to 1.
}

\examples{
scrape_wahlrecht_bund()
}
