% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAlts.R
\name{setAlts}
\alias{setAlts}
\title{setAlts
Properly formats a user supplied a set of alternatives.}
\usage{
setAlts(matrixOfAlts = NA)
}
\arguments{
\item{matrixOfAlts}{a numberOfALts x numOfDimensions matrix containing the location of the alternatives.}
}
\value{
outAlternativeDataFrame
}
\description{
Takes a set of alternatives and puts them in the proper format. 
The user should supply the alternatives as
a numberOfALts x numOfDimensions matrix.
}
\examples{
  setAlts(matrixOfAlts = cbind( c(0), c(0) ) )
  setAlts(matrixOfAlts = cbind( c(.1,.2,.3,.4), c(.5,.6,.7,.8) ) )
}
