% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCompetitors.R
\name{setCompetitors}
\alias{setCompetitors}
\title{setCompetitors
Properly formats a user supplied a set of competitors.}
\usage{
setCompetitors(matrixOfCompetitors = NA)
}
\arguments{
\item{matrixOfCompetitors}{a numberOfCompetitors x numOfDimensions matrix containing the location of the competitors}
}
\value{
outCompetitorDataFrame
}
\description{
Takes a set of competitors and puts them in the proper format for voteR to use. The user should supply the competitors as
a numberOfCompetitors x numOfDimensions matrix.
}
\examples{
  setCompetitors(matrixOfCompetitors = cbind( c(.1,.2,.3,.4), c(.5,.6,.7,.8) ) )
}
