% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{sum}
\alias{sum}
\title{Sum of vector elements.}
\usage{
sum(..., na.rm = TRUE)
}
\description{
\code{sum} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}
