% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_run.R
\name{summarise_RUN}
\alias{summarise_RUN}
\title{Summarize_Run}
\usage{
summarise_RUN(run = import_run())
}
\arguments{
\item{run}{a run data frame returned by \code{import_run}.}
}
\value{
A ggplot object.
}
\description{
Get summary statistics for run
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
run2profile(smooth = 0.5)
}
