% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useApprovalVoting.R
\name{useApprovalVoting}
\alias{useApprovalVoting}
\title{useApprovalVoting
Applies the COndorcet Method to a set of preferences. Ties broken at random.}
\usage{
useApprovalVoting(
  voterPreferencesApprovalVoting,
  breakVoterTiesApprovalVoting = "random",
  forceUniqueChoiceApprovalVoting = TRUE,
  howToSelectUniqueChoiceApprovalVoting = "random"
)
}
\arguments{
\item{voterPreferencesApprovalVoting}{A matrix of voter preferences that is numVoters x numAlts.}

\item{breakVoterTiesApprovalVoting}{Determines how a voter with two equal top choices cast their vote. Can be "random", "first" or "last".}

\item{forceUniqueChoiceApprovalVoting}{If there is more than one alternative in the choice set should only one be returned as the gorup choice. If set to TRUE then choose a method with the parameter \code{howToSelectUniqueChoice}.}

\item{howToSelectUniqueChoiceApprovalVoting}{If there is more than one alternative in the choice set, and \code{forceUniqueChoice == "TRUE"}, this determines which alternative is selected as the gorup choice. Can be "random", "first" or "last".}
}
\value{
A list where the first element is the group's choice set and the second element is the vote totals.
}
\description{
Takes in a set of voter preferences, applies plurality rule to them and returns the choice set.
}
