% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePluralityRule.R
\name{usePluralityRule}
\alias{usePluralityRule}
\title{usePluralityRule
Applies plurality rule to a set of preferences. Ties broken at random.}
\usage{
usePluralityRule(
  voterPreferences,
  breakVoterTies = "random",
  forceUniqueChoice = "TRUE",
  howToSelectUniqueChoice = "random"
)
}
\arguments{
\item{voterPreferences}{A matrix of voter preferences that is numVoters x numAlts.}

\item{breakVoterTies}{Determines how a voter with two equal top choices cast their vote. Can be "random", "first" or "last".}

\item{forceUniqueChoice}{If there is more than one alternative in the choice set should only one be returned as the gorup choice. If set to TRUE then choose a method with the parameter \code{howToSelectUniqueChoice}.}

\item{howToSelectUniqueChoice}{If there is more than one alternative in the choice set, and \code{forceUniqueChoice == "TRUE"}, this determines which alternative is selected as the gorup choice. Can be "random", "first" or "last".}
}
\value{
A list where the first element is the group's choice set and the second element is the vote totals.
}
\description{
Takes in a set of voter preferences, applies plurality rule to them and returns the choice set.
}
