% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePluralityRunoffRule.R
\name{usePluralityRunoffRule}
\alias{usePluralityRunoffRule}
\title{usePluralityRunoffRule
Applies plurality rule to a set of preferences. Ties broken at random.}
\usage{
usePluralityRunoffRule(
  voterPreferencesPluralityRunoff,
  breakVoterTiesPluralityRunoff = "random"
)
}
\arguments{
\item{voterPreferencesPluralityRunoff}{A matrix of voter preferences that is numVoters x numAlts.}

\item{breakVoterTiesPluralityRunoff}{Determines how a voter with two equal top choices cast their vote. Can be "random", "first" or "last".}
}
\value{
A list where the first element is the group's choice set and the second element is the vote totals.
}
\description{
Takes in a set of voter preferences, applies plurality rule to them and returns the choice set.
}
