% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSequentialRunoffRule.R
\name{useSequentialRunoffRule}
\alias{useSequentialRunoffRule}
\title{useSequentialRunoffRule
Applies plurality rule to a set of preferences. Ties broken at random.}
\usage{
useSequentialRunoffRule(
  voterPreferencesSequentialRunoff,
  breakVoterTiesSequentialRunoff = "random"
)
}
\arguments{
\item{voterPreferencesSequentialRunoff}{A matrix of voter preferences that is numVoters x numAlts.}

\item{breakVoterTiesSequentialRunoff}{Determines how a voter with two equal top choices cast their vote. Can be "random", "first" or "last".}
}
\value{
A list where the first element is the group's choice set and the subsequent elelmets are the vote toatls form each round.  #STILL TO DO
}
\description{
Takes in a set of voter preferences, applies plurality rule to them and returns the choice set.
}
