% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiTessellation.R
\name{voronoiTessellation}
\alias{voronoiTessellation}
\title{voronoiTessellation
Creates a voroni tessilation of the space.}
\usage{
voronoiTessellation(
  xLocationsVoronoiTessellation,
  yLocationsVoronoiTessellation
)
}
\arguments{
\item{xLocationsVoronoiTessellation}{The x coordinates for the generator points. Usually this will be the alternatives/competitors in the choice set.}

\item{yLocationsVoronoiTessellation}{The y coordinates for the generator points. Usually this will be the aternatives/competitors in the choice set.}
}
\value{
outAlternativeDataFrame
}
\description{
Used mostly for testign purposes. This fucntion allows you to quickly check the cutlines
for 2 alterantive cases and the analogous boundaries for more than 2 alterantives. Only works for 
casses with purely 'circular' euclidian preferences.
}
