% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteByUtilNoCumu.R
\name{voteByUtilNoCumu}
\alias{voteByUtilNoCumu}
\title{voteByUtilNoCumu
Allocate votes based utility values, no cumulation is allowed. 
 
Voters allocate their votes based on their preference prdering as determined by their utility function.
They are not allowd to vote for a competitor or alternative more than once. This function is usually not called directly 
but is wrapped by  \code{\link{votersVote}}.}
\usage{
voteByUtilNoCumu(votersUtilsForAltsVoteByUtilNoCumu,
  numVotesPerVoterVoteByUtilNoCumu)
}
\arguments{
\item{votersUtilsForAltsVoteByUtilNoCumu}{numVoters by numAlts matrix of utility values.}

\item{numVotesPerVoterVoteByUtilNoCumu}{the number of votes each voter is allowed to cast. This function does not allow cumulation of votes.}
}
\value{
outVotesVoteByUtilNoCumu A 1 x numberOfAlts matrix displaying the number of votes cast for each alternative.
}
\description{
voteByUtilNoCumu
Allocate votes based utility values, no cumulation is allowed. 
 
Voters allocate their votes based on their preference prdering as determined by their utility function.
They are not allowd to vote for a competitor or alternative more than once. This function is usually not called directly 
but is wrapped by  \code{\link{votersVote}}.
}
