% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Voter_File_function.R
\name{Voter_file}
\alias{Voter_file}
\title{A Voter File Function}
\usage{
Voter_file(State, filelocation, filename)
}
\arguments{
\item{State}{What state voter file do you have?}

\item{filelocation}{Where are your files held?}

\item{filename}{What is your file called?}
}
\description{
This function allows you to read your voter files
}
\examples{
Voter_file(Florida, ~Documents/TylerVF, FloridaFiles)
}
