% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votersVote.R
\name{votersVote}
\alias{votersVote}
\title{votersVote
Alocates votes for a  set of competitors/alternatives according to a set of model parameters.
 
Voters cast their votes for the competitors or alternatives in the choice set based on a set of model parameters.}
\usage{
votersVote(
  voterDataFrameVotersVote = NA,
  altsDataFrameVotersVote = NA,
  abstentionThresholdVotersVote,
  numVotesPerVoterVotersVote,
  cumulationParameterVotersVote
)
}
\arguments{
\item{voterDataFrameVotersVote}{a voterR voter dataFrame that contains the voterID, ideal point location, minkowski order, loss order, and salience weights for each voter.}

\item{altsDataFrameVotersVote}{a voteR dataFrame that contains the competitorID or alternativeID, and current location of competitors or alternatives being voted on.}

\item{abstentionThresholdVotersVote}{A voter will not cast a vote for any competitor/alternative that is past this euclidian distance.}

\item{numVotesPerVoterVotersVote}{The number of votes a voter may cast in the current electoral system.}

\item{cumulationParameterVotersVote}{The degree to which voters want to cumulate votes. If = 0, cumulation is not allowed, For all values > 0 the loss parameter in the voters loss function is used.}
}
\value{
outVoteTotals A dataFrame containing the vote totals for each alternative or competitor.
}
\description{
votersVote
Alocates votes for a  set of competitors/alternatives according to a set of model parameters.
 
Voters cast their votes for the competitors or alternatives in the choice set based on a set of model parameters.
}
